requirements.txt
Flask==2.2.5
cryptography==41.0.2
python-jose[cryptography]==3.3.0
requests==2.31.0